A window can reach past the loop, and what it reaches into is silence
In and out are signed now. An in point before zero, or an out point past
the length, is rest — up to one loop's worth on either side — so a window
extends a loop as well as trims it, and nothing in the arena moves to make
room: counter-rotate at the start and the loop gains a bar of silence
before its downbeat. Outside the arena `loop_at` reads nothing rather than
something; the seam crossfade follows the loop's own next audio when the
window ends inside it and silence when it already reaches past the end.
Peaks are drawn over the loop plus whatever silence the window reaches
into, and say `from` and `to`. The snapshot's `pos` is signed to match.

The ramp test gains the extension: twenty of rest, the loop where it was,
twenty of rest; and a request past a loop's worth is refused by name.

+++ daemon/src/ws.rs
@@ -184,7 +184,7 @@ fn snapshot(sh: &Shared, sr: u32, alive: bool) -> String {
             // Through the engine's own playhead rather than subtracting `origin`
             // here, so the display cannot disagree with the audio about where a
             // loop is — which it would the moment speed or a pendulum was on.
-            let pos = lp.play_pos(cur, len) as usize;
+            let pos = lp.play_pos(cur, len) as i64;
             let shapes: Vec<String> = (0..lp.n_layers.load(Ordering::Acquire))
                 .map(|l| {
                     let (slen, period, phase) = lp.layer_shape(l);
@@ -328,7 +328,7 @@ fn snapshot(sh: &Shared, sr: u32, alive: bool) -> String {
     let sel = sh.sel();
     let cl = sh.lp(sel);
     let loop_len = cl.loop_len.load(Ordering::Acquire);
-    let pos = cl.play_pos(cur, loop_len) as usize;
+    let pos = cl.play_pos(cur, loop_len) as i64;
 
     // Peaks are swapped out, so each reader gets the peak since the last read
     // rather than a decaying maximum. With one client that is exactly right;
